requirements.txt
pandas==1.3.4
numpy==1.20.3
dash==2.10.2
scikit-learn==0.24.2
pyyaml==6.0
dill==0.3.6